relocates tillage implements enum to data structures dir
