+++ RUFAS/data_structures/events.py
@@ -8,9 +8,9 @@ specific harvest method will be used when harvesting a crop, and a `crop_referen
 crop that is presently growing in a field will be harvested.
 """
 
-from RUFAS.routines.EEE.enums import TillageImplement
-from RUFAS.routines.field.crop.harvest_operations import HarvestOperation
 from RUFAS.data_structures.manure_types import ManureType
+from RUFAS.data_structures.tillage_implements import TillageImplement
+from RUFAS.routines.field.crop.harvest_operations import HarvestOperation
 from RUFAS.time import Time
