+++ RUFAS/routines/field/field/tillage_application.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
+from RUFAS.data_structures.tillage_implements import TillageImplement
 from RUFAS.output_manager import OutputManager
-from RUFAS.routines.EEE.enums import TillageImplement
 from RUFAS.routines.field.field.field_data import FieldData
 from RUFAS.routines.field.soil.manure_pool import ManurePool
 from RUFAS.routines.field.soil.soil_data import SoilData
