+++ RUFAS/routines/field/manager/tillage_schedule.py
@@ -1,7 +1,7 @@
 from typing import List, Any
 
-from RUFAS.routines.EEE.enums import TillageImplement
 from RUFAS.data_structures.events import TillageEvent
+from RUFAS.data_structures.tillage_implements import TillageImplement
 from RUFAS.routines.field.manager.schedule import Schedule
 from RUFAS.util import Utility
