+++ tests/soil_crop_tests/manager_tests/test_events.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from RUFAS.routines.EEE.enums import TillageImplement
+from RUFAS.data_structures.tillage_implements import TillageImplement
 from RUFAS.routines.field.crop.harvest_operations import HarvestOperation
 from RUFAS.data_structures.events import (
     BaseFieldManagementEvent,
